requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.5
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-chartjs==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
future==0.18.2
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2020.5
reportlab==3.5.65
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
xhtml2pdf==0.2.5
